Add logic for aTruncated to display
truncate() runs in updateScreen(), evaluating the length of a. If a is long enough, aTruncated is displayed in place of a, but a maintains its value in calculations. The full value will be displayed in the log.

operate() changes a and b into numbers before any operation.

Enter key default action is prevented, stopping focused buttons from being clicked (ex. if calculating with Shift + Enter and focus is on "C", the calculation would occur with "C" pressed via Enter).